+++ src/chevah_compat/tests/elevated/test_filesystem.py
@@ -7,7 +7,12 @@ Tests for portable filesystem access.
 import errno
 import os
 
-from chevah_compat import LocalFilesystem, SuperAvatar
+from chevah_compat import (
+    DefaultAvatar,
+    LocalFilesystem,
+    SuperAvatar,
+    system_users,
+)
 from chevah_compat.exceptions import CompatError, CompatException
 from chevah_compat.interfaces import IFileAttributes
 from chevah_compat.testing import (
@@ -310,6 +315,108 @@ class TestPosixFilesystem(FileSystemTestCase):
         self.assertAlmostEqual(self.now(), file_attributes.modified, delta=5)
         self.assertIsInstance(str, file_attributes.name)
 
+    def test_impersonate_nested(self):
+        """
+        The user impersonation works for nested calls.
+
+        Once all nested calls exit, the OS process is reset to the previous
+        user.
+        """
+        user = TEST_USERS['normal']
+        avatar = mk.FilesystemOsAvatar(
+            user=user,
+            home_folder_path=mk.fs.temp_path,
+        )
+
+        initial_user = system_users.getCurrentUserName()
+        filesystem = LocalFilesystem(avatar=avatar)
+        with filesystem._impersonateUser():
+            self.assertEqual(user.name, system_users.getCurrentUserName())
+            with filesystem._impersonateUser():
+                self.assertEqual(user.name, system_users.getCurrentUserName())
+
+            # On Windows, nested calls are not supported.
+            # I don't know how to implement nested support in Windows.
+            if self.os_family != 'nt':
+                # Even though we have exited the context,
+                # we still have the impersonated use as this is a nested call.
+                self.assertEqual(user.name, system_users.getCurrentUserName())
+
+        # Once we exit all context, the previous context is set.
+        self.assertEqual(initial_user, system_users.getCurrentUserName())
+
+    def test_nested_no_reset(self):
+        """
+        The user impersonation is not reset when nesting specific user
+        filesystem with the default filesystem
+        """
+        user = TEST_USERS['normal']
+        avatar = mk.FilesystemOsAvatar(
+            user=user,
+            home_folder_path=mk.fs.temp_path,
+        )
+        initial_user = system_users.getCurrentUserName()
+        filesystem = LocalFilesystem(avatar=avatar)
+        default_filesystem = LocalFilesystem(avatar=DefaultAvatar())
+
+        with default_filesystem._impersonateUser():
+            # Previous user is kept
+            self.assertEqual(initial_user, system_users.getCurrentUserName())
+
+        with filesystem._impersonateUser():
+            self.assertEqual(user.name, system_users.getCurrentUserName())
+
+            with default_filesystem._impersonateUser():
+                # Still uses the nested user.
+                self.assertEqual(user.name, system_users.getCurrentUserName())
+
+        # Once we exit all context, the previous context is set.
+        self.assertEqual(initial_user, system_users.getCurrentUserName())
+
+    def test_nested_with_reset(self):
+        """
+        The user impersonation can reset when nesting a specific user
+        filesystem with the default filesystem.
+        """
+        user = TEST_USERS['normal']
+        avatar = mk.FilesystemOsAvatar(
+            user=user,
+            home_folder_path=mk.fs.temp_path,
+        )
+        initial_user = system_users.getCurrentUserName()
+        filesystem = LocalFilesystem(avatar=avatar)
+
+        default_filesystem = LocalFilesystem(avatar=DefaultAvatar())
+
+        initial_context = DefaultAvatar._NoOpContext
+        DefaultAvatar.setupResetEffectivePrivileges()
+
+        def revert_avatar(initial_context):
+            DefaultAvatar._NoOpContext = initial_context
+
+        self.addCleanup(revert_avatar, initial_context)
+
+        with default_filesystem._impersonateUser():
+            # Previous user is kept
+            self.assertEqual(initial_user, system_users.getCurrentUserName())
+
+        with filesystem._impersonateUser():
+            self.assertEqual(user.name, system_users.getCurrentUserName())
+
+            with default_filesystem._impersonateUser():
+                # Reset the user.
+                self.assertEqual(
+                    initial_user, system_users.getCurrentUserName()
+                )
+
+            # Further call to the specific impersonated filesystem will work
+            # and will trigger an impersonation.
+            with filesystem._impersonateUser():
+                self.assertEqual(user.name, system_users.getCurrentUserName())
+
+        # Once we exit all context, the previous context is set.
+        self.assertEqual(initial_user, system_users.getCurrentUserName())
+
 
 @conditionals.onOSFamily('posix')
 class TestUnixFilesystem(FileSystemTestCase):
@@ -328,7 +435,7 @@ class TestUnixFilesystem(FileSystemTestCase):
     def test_addGroup_denied_group_file(self):
         """
         On Unix we can not set the group for a file that we own to a group
-        to which we are not members, with the exception of HPUX.
+        to which we are not members.
         """
         file_name = mk.makeFilename()
         file_segments = self.filesystem.home_segments
@@ -339,24 +446,15 @@ class TestUnixFilesystem(FileSystemTestCase):
         def act():
             self.filesystem.addGroup(file_segments, TEST_ACCOUNT_GROUP_OTHER)
 
-        if self.os_name == 'hpux':
+        with self.assertRaises(CompatError) as context:
             act()
-            self.assertTrue(
-                self.filesystem.hasGroup(
-                    file_segments,
-                    TEST_ACCOUNT_GROUP_OTHER,
-                ),
-            )
-        else:
-            with self.assertRaises(CompatError) as context:
-                act()
-            self.assertEqual(1017, context.exception.event_id)
-            self.assertFalse(
-                self.filesystem.hasGroup(
-                    file_segments,
-                    TEST_ACCOUNT_GROUP_OTHER,
-                ),
-            )
+        self.assertEqual(1017, context.exception.event_id)
+        self.assertFalse(
+            self.filesystem.hasGroup(
+                file_segments,
+                TEST_ACCOUNT_GROUP_OTHER,
+            ),
+        )
 
     def test_addGroup_denied_group_folder(self):
         """
@@ -371,24 +469,15 @@ class TestUnixFilesystem(FileSystemTestCase):
         def act():
             self.filesystem.addGroup(folder_segments, TEST_ACCOUNT_GROUP_OTHER)
 
-        if self.os_name == 'hpux':
+        with self.assertRaises(CompatError) as context:
             act()
-            self.assertTrue(
-                self.filesystem.hasGroup(
-                    folder_segments,
-                    TEST_ACCOUNT_GROUP_OTHER,
-                ),
-            )
-        else:
-            with self.assertRaises(CompatError) as context:
-                act()
-            self.assertEqual(1017, context.exception.event_id)
-            self.assertFalse(
-                self.filesystem.hasGroup(
-                    folder_segments,
-                    TEST_ACCOUNT_GROUP_OTHER,
-                ),
-            )
+        self.assertEqual(1017, context.exception.event_id)
+        self.assertFalse(
+            self.filesystem.hasGroup(
+                folder_segments,
+                TEST_ACCOUNT_GROUP_OTHER,
+            ),
+        )
 
     def test_removeGroup(self):
         """